Moved other scripts out of contents folder
