+++ components/score.tsx
@@ -5,7 +5,13 @@ export const Score = (props: { data: BrandData }): JSX.Element => {
     <>
       <h1>Goji Score</h1>
       <h2>{props.data.name}</h2>
-      <p>{props.data.categoryValues["Carbon Emissions"]}</p>
+      {props.data.categoryValues.forEach((value, category, _) => {
+        return (
+          <p>
+            {category}: {value}
+          </p>
+        )
+      })}
     </>
   )
 }
